requirements.txt
flask==0.12.2
flask-restful==0.3.6
flask-restplus==0.10.1
flask-bootstrap==3.3.7.1
flask-testing==0.7.1
tinydb==3.8.0
wheel==0.30.0
RPi.GPIO==0.6.3
attrs==17.4.0
cattrs==0.6.0
typing==3.6.4
marshmallow==3.0.0b7